Moved mode from Camera to Base Station
